Move AddProperty to popup, organize files
BREAKING: reroute home to /properties
