+++ src/App.tsx
@@ -1,8 +1,8 @@
-import { Routes, Route } from 'react-router-dom';
+import { Routes, Route, Navigate } from 'react-router-dom';
 import styled from 'styled-components';
 import Navbar from './components/Navbar';
 import Home from './pages/Home';
-import AddProperty from './pages/AddProperty';
+import AddProperty from './components/properties/AddProperty';
 import Tasks from './pages/Tasks';
 
 
@@ -13,8 +13,11 @@ function App() {
       <Navbar />
       <MainContainer>
         <Routes>
-          <Route path="/" element={<Home />} />
-          <Route path="/add-property" element={<AddProperty />} />
+          <Route
+            path="/"
+            element={<Navigate to="/properties" replace />}
+          />
+          <Route path="/properties/*" element={<Home />} />
           <Route path="/property/:id/*" element={<Tasks />} />
         </Routes>
       </MainContainer>
